yaml_schema/requirements.txt
contextlib2==0.6.0.post1
PyYAML==5.3.1
schema==0.7.3
